imported css transition group and reorganize header and footer into folders
