+++ package.json
@@ -14,6 +14,7 @@
     "react-loader-spinner": "^4.0.0",
     "react-router-dom": "5.3.0",
     "react-scripts": "4.0.3",
+    "react-transition-group": "^4.4.2",
     "styled-components": "^5.3.3",
     "web-vitals": "^1.1.2"
   },
